+++ apps/web/src/app/(desktop-ready)/recommend/components/RecommendPageHeader.tsx
@@ -1,15 +1,16 @@
 'use client';
 
-import { Search, Share } from '@/components/common/icons';
+import { Search } from '@/components/common/icons';
 import { useToast } from '@/components/common/Toast';
 import BackButton from '@/components/layout/BackButton';
+import ShareButton from '@/components/ShareButton';
 import { PAGE } from '@/constants/page';
 import Link from '@/features/Link';
 
 import useRecommendedKeyword from '../hooks/useRecommendedKeyword';
 
 export default function RecommendPageHeader() {
-  const { recommendedKeyword } = useRecommendedKeyword();
+  const { recommendedKeyword: keyword } = useRecommendedKeyword();
 
   const { toast } = useToast();
 
@@ -21,35 +22,7 @@ export default function RecommendPageHeader() {
     // });
   };
 
-  const handleShare = () => {
-    if (navigator.share) {
-      // TODO: Need GTM Migration
-      // mp?.track(EVENT.PRODUCT_SHARE.NAME, {
-      //   type: EVENT.PRODUCT_SHARE.TYPE.SHARE_API,
-      //   page: EVENT.PAGE.RECOMMEND,
-      // });
-
-      navigator.share({
-        title: '지금 추천하는 상품',
-        url: window.location.href,
-      });
-    } else {
-      // TODO: Need GTM Migration
-      // mp?.track(EVENT.PRODUCT_SHARE.NAME, {
-      //   type: EVENT.PRODUCT_SHARE.TYPE.NOT_SHARE_API,
-      //   page: EVENT.PAGE.RECOMMEND,
-      // });
-
-      navigator.clipboard.writeText(window.location.href);
-      toast(
-        <>
-          링크가 클립보드에 복사되었어요!
-          <br />
-          원하는 곳에 붙여 넣어 공유해보세요!
-        </>,
-      );
-    }
-  };
+  const title = keyword ? `${keyword} 추천 상품 | 지름알림` : '지금 추천하는 상품';
 
   return (
     <header className="fixed top-0 z-50 flex h-[56px] w-full max-w-screen-mobile-max items-center justify-between bg-white px-5">
@@ -67,14 +40,7 @@ export default function RecommendPageHeader() {
         >
           <Search />
         </Link>
-        <button
-          onClick={handleShare}
-          className="-m-2 p-2 hover:cursor-pointer"
-          aria-label="공유하기"
-          title="공유하기"
-        >
-          <Share />
-        </button>
+        <ShareButton title={title} />
       </div>
     </header>
   );
