+++ apps/web/src/app/(mobile)/products/[id]/components/ProductDeatilPageHeader.tsx
@@ -3,18 +3,15 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { notFound } from 'next/navigation';
 
-import { Search, Share } from '@/components/common/icons';
-import { useToast } from '@/components/common/Toast';
+import { Search } from '@/components/common/icons';
+import LogoLink from '@/components/common/Logo/LogoLink';
 import BackButton from '@/components/layout/BackButton';
+import ShareButton from '@/components/ShareButton';
 import { PAGE } from '@/constants/page';
 import { ProductQueries } from '@/entities/product';
 import Link from '@/features/Link';
 
-import LogoLink from '../../../../../components/common/Logo/LogoLink';
-
 export default function ProductDetailPageHeader({ productId }: { productId: number }) {
-  const { toast } = useToast();
-
   const {
     data: { product },
   } = useSuspenseQuery(ProductQueries.product({ id: productId }));
@@ -29,36 +26,6 @@ export default function ProductDetailPageHeader({ productId }: { productId: numb
     //   type: EVENT.PRODUCT_SEARCH.TYPE.CLICK,
     //   page: EVENT.PAGE.DETAIL,
     // });
-  };
-
-  const handleShare = () => {
-    if (navigator.share) {
-      // TODO: Need GTM Migration
-      // mp?.track(EVENT.PRODUCT_SHARE.NAME, {
-      //   type: EVENT.PRODUCT_SHARE.TYPE.SHARE_API,
-      //   page: EVENT.PAGE.DETAIL,
-      // });
-
-      navigator.share({
-        title,
-        url: window.location.href,
-      });
-    } else {
-      // TODO: Need GTM Migration
-      // mp?.track(EVENT.PRODUCT_SHARE.NAME, {
-      //   type: EVENT.PRODUCT_SHARE.TYPE.NOT_SHARE_API,
-      //   page: EVENT.PAGE.DETAIL,
-      // });
-
-      navigator.clipboard.writeText(window.location.href);
-      toast(
-        <>
-          링크가 클립보드에 복사되었어요!
-          <br />
-          원하는 곳에 붙여 넣어 공유해보세요!
-        </>,
-      );
-    }
   };
 
   return (
@@ -77,14 +44,7 @@ export default function ProductDetailPageHeader({ productId }: { productId: numb
         >
           <Search />
         </Link>
-        <button
-          onClick={handleShare}
-          className="-m-2 p-2 hover:cursor-pointer"
-          aria-label="공유하기"
-          title="공유하기"
-        >
-          <Share />
-        </button>
+        <ShareButton title={title} page="DETAIL" />
       </div>
     </header>
   );
